+++ devicemodel/include/vhost.h
@@ -132,19 +132,15 @@ int vhost_dev_start(struct vhost_dev *vdev);
 int vhost_dev_stop(struct vhost_dev *vdev);
 
 /**
- * @brief set backend fd of vhost net.
+ * @brief vhost kernel dev ioctrl function
  *
- * This interface is called to set the backend fd (for example tap fd)
- * to vhost.
+ * This interface is used to operation the vhost dev kernel.
  *
- * @param vdev Pointer to struct vhost_dev.
- * @param backend_fd fd of backend (for example tap fd).
+ * @param vdev Pointer to struct vhost_dev
+ * @param The request to vhost kernel
+ * @param The arguments of vhost kernel operation
  *
- * @return 0 on success and -1 on failure.
+ * @return 0 on success and -1 on failure
  */
-int vhost_net_set_backend(struct vhost_dev *vdev, int backend_fd);
-
-/**
- * @}
- */
+int vhost_kernel_ioctl(struct vhost_dev *vdev, unsigned long int request, void *arg);
 #endif
